possible loss of accuracy in convolve

+++ convolveExpression.cpp
@@ -29,8 +29,8 @@ Variable convolveExpression::evaluate()
 		for (int cy = 0; cy < image->getHeight(); cy++)
 		{
 			Colour p = image->getPixel(cx, cy);
-			x = cx;
-			y = cy;
+			x = (float)cx;
+			y = (float)cy;
 			
 			if (store->mask->getValue(cx, cy) > 0)
 			{
@@ -40,7 +40,7 @@ Variable convolveExpression::evaluate()
 				r = p.r();
 				g = p.g();
 				b = p.b();
-				int kr, kg, kb;
+				float kr, kg, kb;
 				kr = kg = kb = 0;
 				Variable v = arguments.back()->getResult();
 				Kernel* k = v.get<Kernel>();
@@ -53,9 +53,9 @@ Variable convolveExpression::evaluate()
 						kg += k->get(kx, ky)*q.g();
 						kb += k->get(kx, ky)*q.b();
 					}
-				p.r(clamp<int>(kr, 0, 255));
-				p.g(clamp<int>(kg, 0, 255));
-				p.b(clamp<int>(kb, 0, 255));
+				p.r(clamp<int>((int)kr, 0, 255));
+				p.g(clamp<int>((int)kg, 0, 255));
+				p.b(clamp<int>((int)kb, 0, 255));
 				
 			}
